Refactor progress bar setup
